pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = ""
authors = [
    {name = "Lofienjoyer"}
]
requires-python = "^3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "ipython (>=9.10.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "pillow-heif (>=1.3.0,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
